Refactor submit handling into a separate module
